Integrated the findbugs static analysis tool into the build process so that we can analyze the code and turn up obvious bugs.  A number of bugs reported by this tool were fixed, and some false positives were added to the fb-excl.xml file to be ignored.
Completed an initial implementation of a client/server setup so that the database can be used concurrently from multiple clients.  It's not thread-safe, and there is a lot of refining needed, but at least now it's working.  This required every Command implementation to get the output-stream to use from the current session state, so that stuff is working now.  The SelectCommand also was updated so that a TupleProcessor could be provided, to allow us to send tuples from the server to the client as they are generated.

Implemented some new unit-tests for various classes, the most notable being the FileManager and FilePointer classes.  Now that we are starting to test storage-layer code, hopefully this area of the unit-tests can be grown out rather quickly.


git-svn-id: file:///home/donnie/svn/nanodb/trunk@72 ac87e96d-00ce-4298-91f2-52cadfb428aa