More fixes on the write-ahead logger, and integration with the buffer manager.
Data pages now record the last LSN that affect the data page, which allows us to flush pages from the buffer manager and sync the WAL properly before we do so.

Also added code to load and store the transaction-state file, particularly as part of forcing the WAL so that we can actually report transactions as "committed."

We still need a way to flush certain pages from the buffer manager, and to sync a file through the storage manager.  Once these pieces are in place, that will make most of the remaining issues disappear.